pyproject.toml
[project]
name = "oura-sandbox"
version = "0.1.0"
description = "Development sandbox for Oura Ring API applications"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pandas>=2.0",
    "plotly>=5.0",
    "numpy>=1.26",
    "jupyter>=1.0",
    "nbformat>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
apps = [
    "streamlit>=1.30",
    "statsmodels>=0.14",
    "scipy>=1.12",
]
ai = [
    "anthropic>=0.40",
    "openai>=1.50",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
